# backend/src/docs/openapi.yaml
openapi: 3.0.3
info:
  title: HopeCash API
  description: |
    API REST da plataforma HopeCash (NewHope) — gestão financeira pessoal e familiar,
    offline-first. Todas as entidades sincronizáveis carregam `id` (UUID), `version`,
    `updated_at` e `deleted_at` (soft delete). Erros seguem o envelope
    `{ "error": { "code", "message", "details?" } }`.
  version: 1.0.0
servers:
  - url: /api/v1
security:
  - bearerAuth: []
...
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      security: []
      summary: Cadastro de usuário
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name, email, password]
              properties:
                name: { type: string }
                email: { type: string, format: email }
                password: { type: string, minLength: 8 }
      responses:
        '201': { description: 'Usuário criado; retorna tokens', content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
  /auth/login:
    post:
      tags: [Auth]
      security: []
      summary: Login
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string }
                password: { type: string }
                device_name: { type: string }
      responses:
        '200': { description: Tokens emitidos, content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
        '401': { description: Credenciais inválidas }
  /auth/refresh:
    post:
      tags: [Auth]
      security: []
      summary: Rotaciona o refresh token
      requestBody:
        content:
          application/json:
            schema: { type: object, required: [refresh_token], properties: { refresh_token: { type: string } } }
      responses:
        '200': { description: Novo par de tokens }
...
  /families/{id}/members:
    get:
      tags: [Families]
      summary: Membros da família
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '200': { description: OK } }
  /families/{id}/invites:
    post:
      tags: [Families]
      summary: Convida membro por e-mail
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses: { '201': { description: Convite criado } }
...
  /cards/{cardId}/invoices:
    get: { tags: [Cards], summary: Faturas do cartão, parameters: [{ name: cardId, in: path, required: true, schema: { type: string } }], responses: { '200': { description: OK } } }
    post: { tags: [Cards], summary: Cria fatura, parameters: [{ name: cardId, in: path, required: true, schema: { type: string } }], responses: { '201': { description: Criada } } }
  /cards/{cardId}/invoices/{invoiceId}/pay:
    post:
      tags: [Cards]
      summary: Paga a fatura (gera despesa na conta indicada)
      parameters:
        - { name: cardId, in: path, required: true, schema: { type: string } }
        - { name: invoiceId, in: path, required: true, schema: { type: string } }
      responses: { '200': { description: OK } }
...
  /retaguarda/auth/login:
    post:
      tags: [Retaguarda]
      security: []
      summary: Login de usuário da retaguarda
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        '200': { description: 'Tokens da retaguarda', content: { application/json: { schema: { $ref: '#/components/schemas/AuthTokens' } } } }
...
components:
  securitySchemes:
    bearerAuth: { type: http, scheme: bearer, bearerFormat: JWT }
...
  schemas:
    AuthTokens:
      type: object
      properties:
        data:
          type: object
          properties:
            access_token: { type: string }
            refresh_token: { type: string }
            user: { type: object }
...